% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{codex_client_info}
\alias{codex_client_info}
\title{Inspect a Codex client}
\usage{
codex_client_info(client)
}
\arguments{
\item{client}{An \code{artcodex_client}.}
}
\value{
A serializable list describing process and server state.
}
\description{
Inspect a Codex client
}
