% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{codex_client_is_alive}
\alias{codex_client_is_alive}
\title{Test whether a Codex client process is alive}
\usage{
codex_client_is_alive(client)
}
\arguments{
\item{client}{An \code{artcodex_client}.}
}
\value{
A single logical value.
}
\description{
Test whether a Codex client process is alive
}
