% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{codex_client_stop}
\alias{codex_client_stop}
\title{Stop a Codex client process}
\usage{
codex_client_stop(client, timeout = 2)
}
\arguments{
\item{client}{An \code{artcodex_client}.}

\item{timeout}{Seconds to allow for a clean exit before terminating it.}
}
\value{
The client, invisibly.
}
\description{
Stop a Codex client process
}
