% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{codex_succeeded}
\alias{codex_succeeded}
\title{Test whether a result completed successfully}
\usage{
codex_succeeded(x)
}
\arguments{
\item{x}{An \code{artcodex_result}.}
}
\value{
A single logical value.
}
\description{
Test whether a result completed successfully
}
