% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{codex_thread_archive}
\alias{codex_thread_archive}
\title{Archive a Codex thread}
\usage{
codex_thread_archive(thread, timeout = 30)
}
\arguments{
\item{thread}{An \code{artcodex_thread}.}

\item{timeout}{Maximum number of seconds to wait.}
}
\value{
The server response, invisibly.
}
\description{
Archive a Codex thread
}
