% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thread.R
\name{codex_turn_interrupt}
\alias{codex_turn_interrupt}
\title{Interrupt a running Codex turn}
\usage{
codex_turn_interrupt(thread, turn_id, timeout = 10)
}
\arguments{
\item{thread}{An \code{artcodex_thread}.}

\item{turn_id}{Identifier of the running turn.}

\item{timeout}{Maximum number of seconds to wait.}
}
\value{
The server response, invisibly.
}
\description{
Interrupt a running Codex turn
}
